MLK-13840 fbdev: mxsfb: Fix explicit null dereference
In mxsfb driver, function overlayfb_check_var, a null pointer dereference
occurs if fourcc pixel format is one not considered explicitly.

This case should not occur, since the pixel format is verified against
supported values before getting to this null dereference code, but this may
change if overlay_fmt_support() or overlayfb_check_var() changes and they
are not kept in sync.

Signed-off-by: Cristina Ciocan <cristina-mihaela.ciocan@nxp.com>