zero-sugar dts: fix OTG regulator config, adjust digitizer and wifi pinmux
OTG power is only enabled on OTG1 (POGO), and is enabled by setting a bit in
a register in the charger, not by toggling GPIO. The GPIO control config in
the OTG regulator blocks are thus removed.

Move digitizer pin muxing from standard pinux block to the lpsr pinmux block.
GPIO1.0..GPIO1.7 requires mux config in separate mux control register, which
then requires the pin muxing to be defined in a new lpsr pinmux block.

Move wifi pinmux to be configured at boot and not when the brcmfmac driver
is loaded (attempt to resolve device init issues).