input: wacom_i2c: drop reset sequence on GPIO
The reset sequence via GPIO makes digitizer non-functional. This was not
an issue when we have GPIO1_0 pad configured as WDOG function, because
the reset sequence setting on the GPIO doesn't take any effect.  But now
pad configuration is gone and the pad works in GPIO mode by default, so
the reset sequence is taking effect.  Let drop this problematic reset
and get digitizer back to work.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>