dts/zero-sugar: remove unreasonable charger settings
We do not really want to disable fast-charge and top-off timer, as the
former prevents charger from staying in fast-charge stage for too long,
and the latter triggers the entering to top-off stage.  So before we
get best settings for our battery pack from vendor, let's drop the
settings from device tree and let driver use the default.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>