dts/zero-sugar: enable keep-power-in-suspend for wifi support
The brcmfmac driver actively tries to set MMC_PM_KEEP_POWER flag in
brcmf_ops_sdio_suspend(). But the call sdio_set_host_pm_flags() fails
to do so, because the flag is missing from host->pm_caps.  To have this
flag set in host->pm_caps, we need to have 'keep-power-in-suspend'
property for usdhc node that wifi is using.

This fixes the issue that wifi stops working after a suspend/resume
cycle.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>