dts/zero-sugar: use fixed regulator to manage WIFI_PWR_EN
Besides WIFI_REG_ON handled by wifi_pwrseq, we need to deal with
WIFI_PWR_EN as well to get BCM43455 WiFi chip powered properly.  This
WIFI_PWR_EN is currently handled in U-Boot.  As we have fixed-regulator
infrastructure in kernel which can handle this nicely, let's use it, so
that we can drop the code from U-Boot.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>