dts/zero-sugar: add pinctrl for bd71815 device
BD71815 device uses GPIO6_IO16 as the PMIC_INT_B interrupt line.  Even
though we are currently not using the interrupt from the device, there
should be a pinctrl for PMIC_INT_B to make the hardware description more
complete and correct.

While at it, rename the device node to bd71815 for an explicit naming.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>