mfd: bd7181x: set LPSR_MODE bit only when system sleeps
Bit LPSR_MODE should be cleared only when system is going to sleep in
LPSR state.  Other than that, the bit should be set for system to get
powered off in SNVS state.

All BUCKs and LDOs will be turned off in SNVS state, only except LDO3.
So we need to change the default setting to turn off LDO3 in SNVS state.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>