max77818-charger: add pm_ops in order to switch/reset pinctrl state after LPSR
When going into LPSR sleep, the power is taken from the GPIO bank used for the
charge state GPIOs.

When returning from LPSR sleep, the GPIO pinctrl state has to be re-initialized
in order to enable the internal pull-up.

As the GPIOs are used as inputs and no other initialization is done during the
initial probe/init of the driver, no further action is required.

[Grzegorz: Fix compilation by including <linux/pinctrl/consumer.h> which
	   is required after porting this patch from kernel v4.14 to v5.4.]
Signed-off-by: Grzegorz Jaszczyk <grzegorz.jaszczyk@linaro.org>