LF-1256-1 drm/imx: imx-ldb: Allocate imx_ldb in ->probe() instead of ->bind()
The structure imx_ldb essentially contains drm encoder
and drm connector.  If we allocate the instance of structure
imx_ldb with devm_kzalloc() in ->bind(), it would be freed after
drm_bridge_attach() or drm_panel_attach() fail.  However,
drm_encoder_init() or drm_connector_init_with_ddc() may
have already added the encoder or the connector into the
encoder_list or the connector_list.  The function
drm_mode_config_cleanup() called in imx_drm_bind() as a part
of it's bailout logic would walk through the lists to try to
cleanup the encoder and the connector, which would cause the
use-after-free issue.  This patch changes the imx_ldb allocation
from ->bind() to ->probe() to fix the issue.  And, with this
change, ->unbind() doesn't need to cleanup the encoder and
the connector anymore, since it can be done by the imx-drm core
driver.

Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>