video: mxsfb: Don't set blank state if enable failed
Check if lcd interface enable was a success before
setting current blank state. This fixes a bug where
the pan function failes when waiting for an interrupt,
when the driver believes the display is unblanked when
in fact it is not.

Also user space applications that call ioctl to do
blanking will be notified when the blanking failed.