sy7636a regulator: Add mutex lock for is_enabled
This prevents a screen freeze bug.

If the regulator checks the on-off bit at the same
time as the thermal driver, the regulator_enable
call will conclude that the regulator is already
active.

Then, when the thermal driver is done checking the
temperature, it will set the PMIC back to the state
it was when it started, which is off.

In this state, the regulator driver believes the
PMIC is on, and the thermal has turned it off.

This can be considered a bit of a quick fix, to make
minimal changes to the kernel at this point. The
correct way to solve this would be to make the
sy7636a_thermal driver use the regulator network, or
moving all register access code into the sy7636a mfd
and use a mutex from there.