max77818-charger: force false chgin connection status
The max77818 charger device reports a valid connection status for
the chgin interface, regardless of whether it has been disabled or not
with chgin_sel.

The charging status is thus not reported correct, as the device is not
charging from the pogo interface even if the connection state indicates
otherweiz.

In order to report a correct state to userspace, the charger connection
state is now always reported to be false for the chgin interface (as
this is always turned off as soon as the connection state changes)